+++ ydb/public/sdk/cpp/src/client/topic/ut/ut_utils/ya.make
@@ -3,6 +3,8 @@ LIBRARY()
 SRCS(
     topic_sdk_test_setup.cpp
     topic_sdk_test_setup.h
+    txusage_fixture.cpp
+    txusage_fixture.h
 )
 
 PEERDIR(
